ijin

+++ src/bridging/DlgDataTB.java
@@ -10,36 +10,19 @@
  */
 
 package bridging;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import fungsi.WarnaTable;
-import fungsi.akses;
-import fungsi.batasInput;
-import fungsi.koneksiDB;
-import fungsi.sekuel;
-import fungsi.validasi;
-import java.awt.Cursor;
-import java.awt.Dimension;
-import java.awt.event.KeyEvent;
-import java.awt.event.KeyListener;
-import java.awt.event.WindowEvent;
-import java.awt.event.WindowListener;
-import java.io.FileReader;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.util.HashMap;
-import java.util.Map;
-import javax.swing.JOptionPane;
-import javax.swing.JTable;
-import javax.swing.event.DocumentEvent;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.TableColumn;
-import laporan.DlgCariPenyakit;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.web.client.RestTemplate;
+import com.fasterxml.jackson.databind.*;
+import fungsi.*;
+import java.awt.*;
+import java.awt.event.*;
+import java.io.*;
+import java.sql.*;
+import java.util.*;
+import javax.swing.*;
+import javax.swing.event.*;
+import javax.swing.table.*;
+import laporan.*;
+import org.springframework.http.*;
+import org.springframework.web.client.*;
 
 /**
  *
@@ -228,8 +211,8 @@ public class DlgDataTB extends javax.swing.JDialog {
         KdKel.setDocument(new batasInput((byte)10).getOnlyAngka(KdKel));
         KeteranganRujukan.setDocument(new batasInput((byte)100).getKata(KeteranganRujukan));
        // PaduanOAT.setDocument(new batasInput((int)500).getKata(PaduanOAT));
-        KeteranganSO.setDocument(new batasInput((int)500).getKata(KeteranganSO));
-        Keterangan.setDocument(new batasInput((int)100).getKata(Keterangan));
+        KeteranganSO.setDocument(new batasInput(500).getKata(KeteranganSO));
+        Keterangan.setDocument(new batasInput(100).getKata(Keterangan));
         PemeriksaanLaboratAkhirNoReg.setDocument(new batasInput((byte)15).getKata(PemeriksaanLaboratAkhirNoReg));
         PemeriksaanLaboratBulan2NoReg.setDocument(new batasInput((byte)15).getKata(PemeriksaanLaboratBulan2NoReg));
         PemeriksaanLaboratBulan3NoReg.setDocument(new batasInput((byte)15).getKata(PemeriksaanLaboratBulan3NoReg));
@@ -3137,6 +3120,9 @@ public class DlgDataTB extends javax.swing.JDialog {
         LCount.setText(""+tabMode.getRowCount());
     }
 
+    /**
+     *
+     */
     public void emptTeks() {
         TNoRw.setText("");
         TNoRM.setText("");
@@ -3240,6 +3226,10 @@ public class DlgDataTB extends javax.swing.JDialog {
         BtnPrint.setEnabled(akses.getkemenkes_sitt());
     }
     
+    /**
+     *
+     * @param norawat
+     */
     public void setNoRM(String norawat){
         TabRawat.setSelectedIndex(0);
         TNoRw.setText(norawat);
