+++ src/kepegawaian/DlgBarcode.java
@@ -10,27 +10,14 @@
  */
 
 package kepegawaian;
-import fungsi.WarnaTable;
-import fungsi.akses;
-import fungsi.batasInput;
-import fungsi.koneksiDB;
-import fungsi.sekuel;
-import fungsi.validasi;
-import java.awt.Cursor;
-import java.awt.Dimension;
-import java.awt.event.KeyEvent;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.Map;
-import javax.swing.JOptionPane;
-import javax.swing.JTable;
-import javax.swing.JTextField;
-import javax.swing.event.DocumentEvent;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.TableColumn;
+import fungsi.*;
+import java.awt.*;
+import java.awt.event.*;
+import java.sql.*;
+import java.util.*;
+import javax.swing.*;
+import javax.swing.event.*;
+import javax.swing.table.*;
 
 /**
  *
@@ -74,9 +61,9 @@ public class DlgBarcode extends javax.swing.JDialog {
         }
         tbBangsal.setDefaultRenderer(Object.class, new WarnaTable());
 
-        Bcd.setDocument(new batasInput((int)25).getKata(Bcd));
+        Bcd.setDocument(new batasInput(25).getKata(Bcd));
         
-        TCari.setDocument(new batasInput((int)100).getKata(TCari));
+        TCari.setDocument(new batasInput(100).getKata(TCari));
         if(koneksiDB.CARICEPAT().equals("aktif")){
             TCari.getDocument().addDocumentListener(new javax.swing.event.DocumentListener(){
                 @Override
@@ -722,6 +709,9 @@ public class DlgBarcode extends javax.swing.JDialog {
         LCount.setText(""+tabMode.getRowCount());
     }
 
+    /**
+     *
+     */
     public void emptTeks() {
         Id.setText("");
         Nik.setText("");
@@ -740,6 +730,10 @@ public class DlgBarcode extends javax.swing.JDialog {
         }  
     }
 
+    /**
+     *
+     * @return
+     */
     public JTextField getTextField(){
         return Id;
     }
@@ -748,7 +742,10 @@ public class DlgBarcode extends javax.swing.JDialog {
         return tbBangsal;
     }
        
-     public void isCek(){
+    /**
+     *
+     */
+    public void isCek(){
         BtnSimpan.setEnabled(akses.getbarcode());
         BtnHapus.setEnabled(akses.getbarcode());
         BtnEdit.setEnabled(akses.getbarcode());
